specify getrandom std feature
getrandom has a semver-breaking change in 0.1.10, where the
`std::error::Error` impl for `getrandom::Error` disappeared for linux,
macos, and other OSes disappeared unless the `std` feature was
specified. specifying `std` as a feature is backwards-compatible with
old getrandom versions, and arguably `std::error::Error` being present
for any target OS regardless of `std` being present was an error in the
first place.